refactor: hoist resolved contacts onto ContactSyncController
Removes the picker's internal loading state. The controller now owns
a resolvedContacts cache (observable) populated after every sync —
including the early-return path when checksum probes report no work
to do. The picker reads from the controller directly; CN-change
updates flow through atomically with no spinner flash on screen.

SendRootScreen gains a .loadingContacts step for the gap between
permission grant and first resolution. In that state the priming
screen stays visible with the primary button replaced by a spinner
via Loadable. ContactsPermissionScreen takes a new isAllowing flag
to drive that.

ResolvedContact, ResolvedContacts, and RecipientLoader move out of
the picker file into a dedicated ContactDirectory.swift so the
controller can depend on them without pulling in the view file.
