+++ Flipcash/Core/Screens/ContactsPermission/ContactsPermissionScreen.swift
@@ -9,12 +9,28 @@ import FlipcashUI
 
 /// Priming view for the system contacts authorization prompt. The parent
 /// owns the ``ContactsAuthorizer`` so grants propagate via `@Observable`.
-/// Pass `onSkipped: nil` to hide the secondary "Not Now" button.
+/// Pass `onSkipped: nil` to hide the secondary "Not Now" button. Set
+/// `isAllowing` to `true` to swap the primary button label for a spinner
+/// — used by parents that need to keep the priming screen visible while
+/// an in-flight grant + downstream load resolves.
 struct ContactsPermissionScreen: View {
 
     let authorizer: ContactsAuthorizer
     let onAllowed: () -> Void
     let onSkipped: (() -> Void)?
+    let isAllowing: Bool
+
+    init(
+        authorizer: ContactsAuthorizer,
+        onAllowed: @escaping () -> Void,
+        onSkipped: (() -> Void)?,
+        isAllowing: Bool = false
+    ) {
+        self.authorizer = authorizer
+        self.onAllowed = onAllowed
+        self.onSkipped = onSkipped
+        self.isAllowing = isAllowing
+    }
 
     // MARK: - Body -
 
@@ -42,12 +58,18 @@ struct ContactsPermissionScreen: View {
 
                 Spacer()
 
-                Button(primaryTitle, action: primaryAction)
-                    .buttonStyle(.filled)
+                Button(action: primaryAction) {
+                    Loadable(isLoading: isAllowing, color: .textAction) {
+                        Text(primaryTitle)
+                    }
+                }
+                .buttonStyle(.filled)
+                .disabled(isAllowing)
 
                 if let onSkipped {
                     Button("Not Now", action: onSkipped)
                         .buttonStyle(.subtle)
+                        .disabled(isAllowing)
                 }
             }
             .padding(20)
