+++ Flipcash/Core/Screens/Main/ScanScreen.swift
@@ -405,10 +405,7 @@ private struct RoutedSheet: View {
             }
         case .send:
             NavigationStack(path: $router[.send]) {
-                SendRootScreen(
-                    container: container,
-                    sessionContainer: sessionContainer
-                )
+                SendRootScreen(container: container)
                 .appRouterDestinations(container: container, sessionContainer: sessionContainer)
                 .navigationTitle("Send")
                 .toolbarTitleDisplayMode(.inline)
